requirements.txt
numpy==1.17.1
Pillow==6.1.0
pydicom==1.2.0
pytz==2018.5
